Add non-vectorized IRQ mode
This patch adds a non-vectorized IRQ mode. System software can
program vector mode via MTVEC[0]. When it is set to 0, CPU fetches
the same entry of MTVEC/MTVECX/UTVEC as the coresponding exceptions
when the IRQs are indicated.

This change allows us to try non-vectorized IRQ based software
without the need to modify it.

Passed core-v-verif hello-world-veri-run, also tested the new IRQ
mode in a VCS simulation environment.

Signed-off-by: Lv Zheng <zhenglv@hotmail.com>